backend/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Dance==7.2.0
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.2.2
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3